Remote_Pi/requirements.txt
requests==1.1.0
python-firebase
wiringpi
pyrebase